+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 
 ### Latest changes
 
+* VFS-13057 Added loading and error handling to "Expose as Public Data..." toggle
 * VFS-12874 Fixed animation on ongoing transfers map
 * VFS-12874 Changed "Share" to "share" naming in GUI texts
 * VFS-12811 Upgraded Ember Bootstrap to 4.9.0